+++ NEWS.md
@@ -25,6 +25,7 @@ exist and fails gracefully with telling error message if yes #4.
 * Replaced `normalizePath()` by `fs::path()` in `p_attribute_encode()` #65.
 * `cwb_get_bindir()` tries to find `cwb-config` system utility, if it is on the path.
 * `cwb_get_url()` will get CWB v3.5 installation files #63.
+* Argument `compress` defaults to `FALSE` as corpus compression is not stable on Windows #3.
 
 cwbtools 0.3.9
 ==============